days/dataflow-framework/file-processing-sys/pyproject.toml
[project]
name = "file-processing-sys"
version = "0.1.0"
description = "A real-time file processing system with FastAPI and CLI."
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "typer",
    "pydantic",
    "rich",
    "networkx",
    "watchdog",
    "pyyaml"
]

[project.scripts]
fpsys = "file_processing_sys.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
